// src/interfaces/results.records.interfaces.ts
export interface APIResponseResultsRecords {
    resultado: Result;
    fichas: string[];
}
export interface hasErrorResult {
    error: string
}
export interface Result {
    img: Array<string[]>;
    fichas: File[] | hasErrorResult;   // o {error: string} si no hay fichas
    caracteristicas: { [key: string]: string[] };
    datos: Datos;
}

export interface Datos {
    codemp: string;
    cantidadFichasEnPag: number;
    cantidadFichas: number;
    cantidad: number;
    guardia: string;
    SiguientePag: number;
    AnteriorPag: number;
    fichasPorPagina: number;
    fichasPorPaginaReal: number;
    paginas: number;
}